--- a/src/Benchmarks/Benchmarks.csproj
+++ b/src/Benchmarks/Benchmarks.csproj
@@ -8,6 +8,7 @@
     <ItemGroup>
       <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
       <ProjectReference Include="..\Humanizer\Humanizer.csproj" />
+<!--      <PackageReference Include="Humanizer" Version="2.14.1" />-->
     </ItemGroup>
 
 </Project>