--- a/src/Benchmarks/EnglishArticleBenchmarks.cs
+++ b/src/Benchmarks/EnglishArticleBenchmarks.cs
@@ -1,0 +1,7 @@
+﻿[MemoryDiagnoser(false)]
+public class EnglishArticleBenchmarks
+{
+    [Benchmark]
+    public string[] AppendArticlePrefix() =>
+        EnglishArticle.AppendArticlePrefix([ "Ant", "The Theater", "The apple", "Fox", "Bear"]);
+}