v2.14.1

| Method              | Mean     | Error     | StdDev    | Allocated |
|-------------------- |---------:|----------:|----------:|----------:|
| AppendArticlePrefix | 3.410 us | 0.0148 us | 0.0139 us |   7.88 KB |

v3.0.0

| Method              | Mean     | Error    | StdDev   | Allocated |
|-------------------- |---------:|---------:|---------:|----------:|
| AppendArticlePrefix | 787.3 ns | 15.45 ns | 20.08 ns |     312 B |